Make P4PythonInstaller work on OS X Mavericks.
OS X Mavericks introduced a warning in Clang that causes builds to fail because
of incorrect arguments to `cc`. The common workaround for this is to use the
ARCHFLAGS environment variable to specify an additional -Wno-error in order to
turn this back into something that doesn't fail, but the P4Python installer
overrides ARCHFLAGS, so instead we override CFLAGS.

Testing done:
Installed from setup.py successfully for the first time since upgrading to
Mavericks.

Reviewed at https://reviews.reviewboard.org/r/5822/