Add a separate 'python setup.py build' step that uses '--ssl'
We originally didn't add SSL support to P4PythonInstaller because the first
iteration of the SSL implementation in it required passing in the library path
for openssl. They've since changed that to make the library path optional, in
which case it will search the system paths.

This change does cause P4PythonInstaller to require openssl libs, but that's a
pretty trivial change in dependencies.

Testing done:
Built and checked that the resulting .so file links against libssl. The libssl
in OS X's default path is 0.9.8, which isn't the right version, so I still get
errors when trying to use ssl: P4PORTs on my devserver, but it's actually
trying to use SSL instead of the stubs.

Also installed this on RBCommons and was able to get further along in the
verification stages with an SSL perforce server.

Reviewed at https://reviews.reviewboard.org/r/5823/